backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiosqlite==0.20.0
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1